# Conflicts:
#	Telegram/Telegram-iOS/en.lproj/Localizable.strings
#	submodules/ChatListUI/Sources/Node/ChatListNodeEntries.swift
#	submodules/ChatListUI/Sources/Node/ChatListStorageInfoItem.swift
#	submodules/TelegramCore/Sources/TelegramEngine/Peers/UpdateBotInfo.swift
#	submodules/TelegramUI/Sources/ChatMessageInteractiveFileNode.swift
#	submodules/TelegramUI/Sources/FetchVideoMediaResource.swift
#	submodules/TelegramUI/Sources/PeerInfo/PeerInfoHeaderNode.swift
#	versions.json

+++ submodules/TelegramUI/Sources/PeerInfoGifPaneNode.swift
@@ -15,6 +15,7 @@ import ListMessageItem
 import ChatMessageInteractiveMediaBadge
 import SoftwareVideo
 import ChatControllerInteraction
+import PeerInfoVisualMediaPaneNode
 
 private final class FrameSequenceThumbnailNode: ASDisplayNode {
     private let context: AccountContext
@@ -373,7 +374,7 @@ private final class VisualMediaItemNode: ASDisplayNode {
                         })
                         
                         if let duration = file.duration {
-                            let durationString = stringForDuration(duration)
+                            let durationString = stringForDuration(Int32(duration))
                             
                             var badgeContent: ChatMessageInteractiveMediaBadgeContent?
                             var mediaDownloadState: ChatMessageInteractiveMediaDownloadState?
